Remove vendor directory from repo
For one, this is the recommended way of managing composer deps.

The real reason however is that committing the vendors does not actually
commit the vendors. When you have dev-dependencies, which are by default
installed from source (e.g. git), then doing a `git add` will not add
the files, because git will not add .git directories to the repo.

The result is an empty folder in vendor/*/* which composer thinks is the
actual package where it is actually empty.

Net result: the autoloader cannot find the dependencies and throws a
class not found error.